fix(support) Clean up namespacing of classes
Also ensured that each class, including support classes enclosed within other classes, is broken out into a separate source file.

[Closes #6]

48 tests, 48 assertions, 0 failures, 0 errors, 0 skips
Coverage: 569 / 569 LOC (100.0%) covered.
RuboCop: 27 files inspected, no offenses detected
Flog: total 285.1, method average 3.4, max 6.7 (Rack::ServiceApiVersioning::BuildRedirectLocationUri#combined_path_str)
Flay total score: 0
Reek: 0 total warnings
