+++ lib/rack/service_api_versioning/encoded_api_version_data.rb
@@ -4,7 +4,9 @@ require 'forwardable'
 
 require 'prolog/dry_types'
 
-require_relative './invalid_base_url_error'
+require_relative './encoded_api_version_data/input_data'
+require_relative './encoded_api_version_data/return_data'
+require_relative './encoded_api_version_data/invalid_base_url_error'
 
 # All(?) Rack code is namespaced within this module.
 module Rack
@@ -52,69 +54,7 @@ module Rack
         raise_invalid_base_url_error original_error
       end
 
-      # Unpacks relevant attributes from passed-in input data
-      class InputData < Dry::Struct::Value
-        attribute :api_version, Types::Strict::Symbol
-        attribute :input_data, Types::Hash
-
-        def base_url
-          version_data[:base_url]
-        end
-
-        def name
-          input_data[:name]
-        end
-
-        def vendor_org
-          content_type_parts[VENDOR_ORG_INDEX]
-        end
-
-        private
-
-        # Index 0 will be `application/vnd`; index 1 the org name (eg, `acme`)
-        VENDOR_ORG_INDEX = 1
-        private_constant :VENDOR_ORG_INDEX
-
-        def content_type_parts
-          version_data[:content_type].split('.')
-        end
-
-        def version_data
-          input_data[:api_versions][api_version]
-        end
-      end # class EncodedApiVersionData::InputData
       private_constant :InputData
-
-      # Immutable, structured data type for returned version data.
-      class ReturnData < Dry::Struct::Value
-        SBU_FMT = %r{\A\w+?://.+?/\z}
-        private_constant :SBU_FMT
-
-        constructor_type :strict_with_defaults
-
-        attribute :api_version, Types::Coercible::String
-        attribute :base_url, Types::Strict::String.constrained(format: SBU_FMT)
-        attribute :name, Types::Strict::String
-        attribute :deprecated, Types::Strict::Bool.default(false)
-        attribute :restricted, Types::Strict::Bool.default(false)
-        attribute :vendor_org, Types::Strict::String
-
-        def content_type
-          content_parts.join('.') + '+json'
-        end
-
-        def to_hash
-          super.merge(content_type: content_type)
-               .reject { |key, _| key == :vendor_org }
-        end
-        alias to_h to_hash
-
-        private
-
-        def content_parts
-          ['application/vnd', vendor_org, name, api_version.to_s]
-        end
-      end # class EncodedApiVersionData::ReturnData
       private_constant :ReturnData
     end # class EncodedApiVersionData
   end
