+++ lib/rack/service_api_versioning/service_component_describer.rb
@@ -3,6 +3,8 @@
 require 'rack'
 require 'rack/response'
 
+require_relative './service_component_describer/report_service_not_found'
+
 # All(?) Rack code is namespaced within this module.
 module Rack
   # Module includes our middleware components for managing service API versions.
@@ -19,36 +21,6 @@ module Rack
     # `service_name`) are supplied *by the AVIDA,* which should know what those
     # "ought" to be.
     class ServiceComponentDescriber
-      # Builds Rack::Result to halt request execution, responding with a 404.
-      class ReportServiceNotFound
-        def self.call(service_name)
-          new(service_name).call
-        end
-
-        def call
-          new_response.finish
-        end
-
-        protected
-
-        def initialize(service_name)
-          @service_name = service_name
-          self
-        end
-
-        private
-
-        attr_reader :service_name
-
-        def body
-          %(Service not found: "#{service_name}")
-        end
-
-        def new_response
-          Rack::Response.new(Array(body), 404)
-        end
-      end # class ServiceComponentDescriber::ReportServiceNotFound
-
       DEFAULT_ENV_KEYS = { result: 'COMPONENT_DESCRIPTION' }.freeze
 
       def initialize(app, repository:, service_name:,
